+++ lib/f_http_client.rb
@@ -11,6 +11,10 @@ require 'f_http_client/cache/key'
 require 'f_http_client/cache/null'
 require 'f_http_client/cache/rails'
 
+require 'f_http_client/logger/default'
+require 'f_http_client/logger/null'
+require 'f_http_client/logger/rails'
+
 module FHTTPClient
   class Error < StandardError; end
   # Your code goes here...